requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
PyPDF2==3.0.1
openai==1.54.0
numpy==1.26.0
scikit-learn==1.3.2
gunicorn==21.2.0
